+++ lib/widgets/add_item_widget.dart
@@ -2,9 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:txdx/input/focus.dart';
+import 'package:txdx/providers/scoped_item_notifier.dart';
 import 'package:txdx/providers/selected_item_provider.dart';
 
 import '../providers/item_notifier_provider.dart';
+import '../providers/settings_provider.dart';
+import '../settings.dart';
 
 class AddItemWidget extends ConsumerWidget {
   AddItemWidget({Key? key}) : super(key: key);
@@ -17,6 +20,32 @@ class AddItemWidget extends ConsumerWidget {
   }
 
   final textController = TextEditingController();
+
+
+  void _setTextDefault(String defaultText) {
+    final suggestion = " $defaultText";
+
+    textController.value = TextEditingValue(
+      text: suggestion,
+      selection: const TextSelection.collapsed(offset: 0),
+    );
+  }
+
+  void _focusWithDefaults(WidgetRef ref) {
+    final autoAddFilter = ref.read(settingsProvider).getBool(settingsAutoAddFilter);
+    if (!autoAddFilter) return;
+
+    final filter = ref.read(itemFilter);
+    if (filter == null || filter.isEmpty) return;
+
+    if (filter == 'due:today') {
+      _setTextDefault('due:today');
+    }
+
+    if (RegExp(r'^[@\+]').hasMatch(filter)) {
+      _setTextDefault(filter);
+    }
+  }
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -32,6 +61,15 @@ class AddItemWidget extends ConsumerWidget {
                   return KeyEventResult.handled;
                 }
                 return KeyEventResult.ignored;
+              },
+              onFocusChange: (focus) {
+                if (focus) {
+                  if (textController.text.isEmpty) {
+                    _focusWithDefaults(ref);
+                  }
+                } else {
+                  textController.clear();
+                }
               },
               child: TextField(
                 focusNode: addNewFocusNode,